// Harken-v2-release2025.08/Harken-v2-release2025.08/packages/backend/src/utils/enums/MessageEnum.ts
export enum SectionEnum {
	ENTER_VALID_ID = 'Enter valid section id',
	SECTION_NOT_FOUND = 'Section not found',
	SECTION_DELETE_FAILED = 'Section delete failed',
	SECTION_DELETE_SUCCESS = 'Section delete success',
	SECTION_UPDATE_FAILED = 'Section update failed',
	SECTION_SAVE_SUCCESS = 'Section save success',
	SECTION_SAVE_FAILED = 'Section save failed',
	TYPE = 'type',
	MAP = 'map',
	AERIAL_MAP = 'aerial_map',
	SALE_APPROACH_COMPS_MAP = 'sale_approach_comps_map',
	COST_APPROACH_COMPS_MAP = 'cost_approach_comps_map',
	APPROACH = 'approach',
	ITEM_NOT_FOUND = 'Item not found',
	ITEM_DELETED_SUCCESSFULLY = 'Item deleted successfully',
	ITEM_DELETE_FAIL = 'Item delete fail',
	ITEM_SAVE_FAILED = 'Item save failed',
	ITEM_SAVE_SUCCESS = 'Item saved successfully',
	ITEM_UPDATE_FAIL = 'Item update failed',
	IMAGE = 'image',
	TEXT_BLOCK = 'text_block',
	SUBSECTION = 'subsection',
	PHOTO_PAGES = 'photo_pages',
	INCOME_COMPARISON = 'income_comparison',
}